requirements.txt
python-dotenv
requests
ta
ccxt
pandas
numpy
